+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This file documents notable changes to the project
 
 - Add Server-sent events utilities
 - Add experimental `log` option to `FetchRouter`
+- Add experimental `expressMiddleware` and `koaMiddleware`
 
 **fixes**